Insert examples into markdown
This uses the format `{example: example-name}` to find examples with the
name 'example-name' and replaces it with a complied jade template named
'templates/code-examples.jade'. This jade template loops over all the
examples found based on their extension (language).

This is only an initial implementation and will require the addition of
syntax highligting and styling of the example block.